Add Secure and SameSite to the cookie
The current cookie has neither Secure neither SameSite; this displays a
warning on Firefox and it will stop working in the future, except if
Secure become true and/or SameSite ∈ { "Lax", "Strict" }. This commit
adds SameSite = "Strict" and possibly Secure=true if the site supports
it.

This commit softly requires MediaWiki 1.36.0 in order to use SameSite
in the patched version of jquery.cookie (commit 22c09860 in MediaWiki),
but it works with previous versions.

Bug: T261093
Change-Id: I39781319728cb184591ea1f35a728c8a5279ab04
(cherry picked from commit e932e37ec4ef39fbfd965fcf97589249450ea4c9)